fixed the roles

+++ events/guild/messageReactionAdd.js
@@ -6,7 +6,7 @@ module.exports = async (client, reaction, user) => {
     try {
         if(reaction.emoji.name == "🗑️") {
             const message = await client.channels.cache.get(reaction.message.channelId).messages.fetch(reaction.message.id);
-            message.member.id === client.user.id && message.content && !message.embeds ? message.delete() : null;
+            message && message.member.id === client.user.id && message.content && !message.embeds ? message.delete() : null;
         }
         if (reaction.emoji.name == "⭐") {
             if (user.bot) return;
